+++ system/config/constants.php
@@ -29,7 +29,7 @@ define('ACE', '1.0.01');
 define('CSS3PIE', '1.0.0');
 define('HIGHLIGHTER', '3.0.83');
 define('HTML5SHIV', '3.6.2');
-define('JQUERY', '1.10.1');
+define('JQUERY', '1.10.2');
 define('JQUERY_UI', '1.10.3');
 define('COLORBOX', '1.4.31');
 define('MEDIAELEMENT', '2.13.1');