+++ system/docs/CHANGELOG.md
@@ -5,6 +5,9 @@ Version 3.2.0 (2013-11-XX)
 --------------------------
 
 ### Fixed
+Add all resize modes to the TinyMCE thumbnail image screen (see #6362).
+
+### Fixed
 Correctly add the parameters to the TinyMCE thumbnail image (see #6361).
 
 ### Fixed