+++ system/docs/CHANGELOG.md
@@ -3,3 +3,6 @@ Contao Open Source CMS changelog
 
 Version 3.3.beta1 (2014-XX-XX)
 ------------------------------
+
+### Fixed
+Always bypass the server cache if there is an auth cookie (see #3184).